P4ADEV-2134 extract from ReceiptService the business logic of create receipt and move it to a dedicated class CreateReceiptService
