+++ src/main/java/it/gov/pagopa/pu/debtpositions/controller/ReceiptControllerImpl.java
@@ -3,7 +3,7 @@ package it.gov.pagopa.pu.debtpositions.controller;
 import it.gov.pagopa.pu.debtpositions.controller.generated.ReceiptApi;
 import it.gov.pagopa.pu.debtpositions.dto.generated.ReceiptDTO;
 import it.gov.pagopa.pu.debtpositions.dto.generated.ReceiptWithAdditionalNodeDataDTO;
-import it.gov.pagopa.pu.debtpositions.service.ReceiptService;
+import it.gov.pagopa.pu.debtpositions.service.create.receipt.CreateReceiptService;
 import it.gov.pagopa.pu.debtpositions.util.SecurityUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,21 +12,21 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class ReceiptControllerImpl implements ReceiptApi {
 
-  private final ReceiptService receiptService;
+  private final CreateReceiptService createReceiptService;
 
-  public ReceiptControllerImpl(ReceiptService receiptService) {
-    this.receiptService = receiptService;
+  public ReceiptControllerImpl(CreateReceiptService createReceiptService) {
+    this.createReceiptService = createReceiptService;
   }
 
   @Override
   public ResponseEntity<ReceiptDTO> createReceipt(ReceiptWithAdditionalNodeDataDTO receiptDTO) {
     String accessToken = SecurityUtils.getAccessToken();
-    ReceiptDTO body = receiptService.createReceipt(receiptDTO, accessToken);
+    ReceiptDTO body = createReceiptService.createReceipt(receiptDTO, accessToken);
     return new ResponseEntity<>(body, HttpStatus.OK);
   }
 
   @Override
   public ResponseEntity<ReceiptDTO> getReceiptDetail(Long receiptId) {
-    return ResponseEntity.ok(receiptService.getReceiptDetail(receiptId));
+    return ResponseEntity.ok(createReceiptService.getReceiptDetail(receiptId));
   }
 }
