+++ src/main/java/it/gov/pagopa/pu/debtpositions/service/ReceiptService.java
@@ -1,9 +1,7 @@
 package it.gov.pagopa.pu.debtpositions.service;
 
 import it.gov.pagopa.pu.debtpositions.dto.generated.ReceiptDTO;
-import it.gov.pagopa.pu.debtpositions.dto.generated.ReceiptWithAdditionalNodeDataDTO;
 
 public interface ReceiptService {
-  ReceiptDTO createReceipt(ReceiptWithAdditionalNodeDataDTO receiptDTO, String accessToken);
   ReceiptDTO getReceiptDetail(Long receiptId);
 }
